Separate redux observer files by entity
