+++ client/reducers.js
@@ -1,0 +1,9 @@
+import { combineReducers } from "redux"
+
+import map from "./components/Map/reducer"
+import lyft from "./reducer"
+
+export default combineReducers({
+  map,
+  lyft
+})
