+++ client/reducers/entities_reducer.js
@@ -1,8 +1,8 @@
-import { combineReducers } from "redux"
-import lyftReducer from "./lyft_reducer"
-import mapReducer from "./map_reducer"
+// import { combineReducers } from "redux"
+// import lyftReducer from "../components"
+// import mapReducer from "./map_reducer"
 
-export default combineReducers({
-  lyft: lyftReducer,
-  map: mapReducer
-})
+// export default combineReducers({
+//   lyft: lyftReducer,
+//   map: mapReducer
+// })
