+++ client/sagas.js
@@ -1,4 +1,4 @@
-import lyftSagas from "./lyft_saga"
+import lyftSagas from "./lyft/saga"
 import mapSagas from "../components/Map/saga"
 import { fork } from "redux-saga/effects"
